+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Documentation inaccuracies for `decode`, `decode2`, and `decode3` in `src/lib.rs`.
 
+### Changed
+* Consolidate `Error::LeadingOnes` into `Error::RedundantEncode` for clarity and consistency.
+
 
 ## [0.2.0] (2025-09-24)
 ### Added